Added support for single-attribute SQLData implementations.
This alleviates a lot of the explicit casting that had to be done for DOMAIN and ENUM.

Our example, where "com.aoindustries.net"."Email" is a DOMAIN (that enforces the same rules as our
com.aoindustries.net.Email class in Java).  The Java type implements SQLData, the PostgreSQL type is a DOMAIN,
both implement the same rules, and conversion between them should be as seemless as possible.

Before:
    "INSERT INTO ticket.\"Ticket\" (...) VALUES (?,?,?,?,?,?,?,?,?::\"com.aoindustries.net\".\"Email\",?,?,?,?,?,?,?,?,?,?)",

After:
    "INSERT INTO ticket.\"Ticket\" (...) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)"