Implement single-attribute type map support for SQLData in `readObject`.
This is related to issue #641. (https://github.com/pgjdbc/pgjdbc/issues/641)

This is complementary to our simpler pull request #1377 (https://github.com/pgjdbc/pgjdbc/pull/1377) that handles the `writeObject`.

This includes type inference for DOMAIN and ENUM types that come back from the server as the base type.  Type inference is used in `readObject(String,Class<?>)` and `readObject(int,Class<?>)`, where the type requested is matched to the type map of the connector.

The inference looks for a direct match first, then search for any inheritance matches.  Thus both providing a means to avoid ambiguity (direct match) and support of the separation of interface from implementation (inheritance match).

When more than one type has been registered for the requested type, an SQLException is thrown due to ambiguity.  Please note, however, that inheritance inference will not conflict with direct inference in this base.  Ambiguity is within the categories of direct or inherited.

Although there is some API changes, they are generally compatible changes.  Please consider prioritizing this as it will greatly assist us in using SQLData for our scalar / non-compound types (DOMAIN and ENUM).

Thank you.