Improving API, naming, and descriptions of the inverted type maps used for type inference.
Change named "inferenceMap" to "invertedMap" to better represent what they are.

Deferred creating of inverted maps until needed.

Removed synchronization on typemap.  Instead synchronizing independently on each inverted maps, giving them a chance to detect a replacement of typemap and recreate their inverted forms.

No longer returning inverted maps, but instead changed the methods to return a `Set<String>` of matching types, or an empty set when there are no types matching the given class.  By not returning a map, the implementation is more free to change as-needed in the future.