Added missing  + 1 on initial hashmap sizing.
Also applied same calculation to the other per-field HashMap.

I know this is off-topic for this pull request.  Maybe this warrants its own tiny pull request, because it is simple but could avoid an occasional rehash where one was not expected (due to (int) cast rounding down).  This is in a part of the code most heavily affecting performance.