Testing the limits of one SQLData type reading/writing nested SQLData attribute.
Currently, writing works correctly, but reading fails for a couple of different reasons.  First, readObject without a type parameter has unbounded recursion between itself and getObject.  Second, readObject with a type parameter suffers the ambiguity of not knowing what to get from the type map.

This is beyond the scope of our current goals of getting simple DOMAIN and ENUM working well for user-defined data types.