Implemented custom types for getObject in CallableStatement.
Because the specifics of how getObject will be called is not known at the time of `executeWithFlags(int)`, I had to hold onto the `ResultSet` and delegate to it for `getObject` methods.

With this result set available, I was able to delegate the other get* methods to it, which is more direct than the old method of getting everything as objects, storing in an array, and casting each object on the way out.

Also with the result set available, I implemented the remaining get* methods through simple delegation to ResultSet.

In order to access the type inference of `PgResultSet.getObjectImpl` while delegating to `ResultSet`, I had to change a few places the be `PgResultSet` instead of just `ResultSet`.

There are some "TODO" notes left where I'm not immediately sure what to do.

All of the existing tests pass locally.  We will work on tests for the funtionality we have updated or added.