Implemented user-defined data types for Arrays.
We are trying to keep the streamlined array creation in PgArray as it has been, while extending it to support custom types.  To this end, we have created an abstraction `ValueAccess`, with implementations that get values from various places, such as `PgResultSet` or values directly available during `PgArray.getArray(...)`.

We then pulled the code from `PgResultSet.getObjectImpl` into a helper class `ValueAccessHelper`, that creates objects from any `ValueAccess` (either from `PgResultSet` or direct values during array creation).

We also moved the management of the type map, along with its inverted forms, into a new class `UdtMap`.  Previously, type inference was only performed based on the typemap of the connection.  With this change, the current typemap is used for inference, whether obtained from the connection, the current `getObject(Map)` invocation, the current `Array.getArray(Map)` invocation, or access to the `ResultSet` obtained from an invocation to `Array.getResultSet(Map)`.

That last point is worth clarifying: When a `ResultSet` comes from an `Array` via `Array.getResultSet(Map)`, calls to `ResultSet.getObject(...)` without a map will use the type map from `Array.getResultSet(Map)`, not the connection's type map.  Same for `Array.getResultSet()` without map - uses the type map associated with the array, not the connection, when not specific map provided to the `ResultSet.getObject(...)` methods.

We have scattered quite a few "TODO" notes around.  We will review these later and, if relevant, create separate, smaller pull requests addressing smaller issues found in the code (such as `Array.getArray(...)` and `Array.getResultSet(...)` with a count of `0` returning all elements.  We're not sure if this is part of the spec.  Plus quite a few places where int overflows can happen without being caught by the current implementation, as well as unsafe casts to (int) on large indexes in `Array.getArray(...)` and `Array.getResultSet(...)` calls.

Another observation is that the `PgResultSet` uses `PSQLState.DATA_TYPE_MISMATCH` when cannot convert types.  Should we use this for custom type conversion problems, too?

We still have a lot of work to do, but want to commit this now so we can start work on any issues in the different builds for different Java and JDBC versions.  All tests pass locally.

Final point for this commit: We could implement this with fewer changes to the code by dispatching calls to `Array.getArray(...)` for custom types through the `Array.getResultSet(...)` implementation.  This would have all object conversion end up going through the existing implementation in `PgResultSet`.  We will modify our implementation of `PgArray` to support both of these tactics (more direct via `ValueAccess` and less direct via `PgResultSet`) and benchmark the results.  This will help guide the decisions on whether the greater codebase changes implied by `ValueAccess` are acceptable for this pull request.  We do not want to change things so much that this pull request becomes untenable.