Sets within inverted maps are no longer unmodifiable.
Since inverted maps might now be generated per-request when custom types are provided in getObject(Map), minimizing the overhead of the inverted map creation by adding the first element as Collections.singleton, then upgrading to HashSet once a second type is added for a given class.

In the common case of a direct match (no inheritance matching needed), and a single type registered for a given class, this will result in the direct inverted map containing only singleton sets and the inheritance map never being created.