Added `Enum` support to `setObject(...)`.
This is a pretty big commit.  Since we can't assume a simple one-to-one relationship between oid and pgType, in order to be able to both call SQLOuput from and create the correct user-defined data types on function return in `CallableStatement`, we are passing the pgType and scale through to the `ParameterList`.

The pgType associated with the parameters is used the return value of `PgParameterMetaData.getParameterTypeName(int)`.

The pgType associated with the parameters is also used, in conjunction with the type map, to return the correct Java class from `PgParameterMetaData.getParameterClassName(int)`.

We are also using this information to provide additional casts within `SimpleParameterList`, which is needed to cast as domain type instead of base type when a user-defined data type is mapped.

We have noted that the assumption of a one-to-one relationship from `oid` back to `pgType` is currently an issue.  We will create a version of `TypeInfo.getPGType(int)` that returns the set of all known representations of the type.  This will give us the necessary information to match types correctly, independent of whether the schema quoted, the name is quoted, and if schema present or on querypath.

The existing code that maps an `oid` back to a single `pgType` will favor a mapped type.  This way, whichever form somebody maps a type (schema, no schema, quoted, not quoted) will be returned as the type name in places like `ParameterMetaData` and `ResultSetMetaData`.

We have left "TODO" notes where scaleOrLength should validate the length of the `InputStream` or `Reader` parameters, per the specification.  Now that the scale is passed through to the parameters, the necessary information is available.  We will not implement this verification on this release.

We noted quite a few more "TODO" of things unfinished or we believe are in error (such as a `CompositeParameterList.setLiteralParameter(...)` calling `subparams.setStringParameter(...)`.  We will dobule-check our understanding of these before asking for review.

We don't want to cause grief within the code, so we do not apply globally, but we have been marking "@Override" where missing on any method we interact with.

`PgPreparedStatement` got a while bunch of `*Impl` methods that take an additional `pgType` parameter (null for the default behavior).  These are necessary so the original `pgType` can remain in effect while `SQLOutput` is writing the object through `PgPreparedStatementSQLOutput`.  We marked the existing methods as `final`, and dispatched to the new `protected` `*Impl` versions of the methods.  Thus anybody who has overridden any of the methods will need to override the `*Impl` version instead, maintaining the current `pgType` as appropriate.  We could have used a `ThreadLocal` to avoid this extra parameter, but we're trying to avoid any "spooky action at a distance" within the code.  By adding the additional parameter it is easy to see what is happening, where from, where to, and why.

We still have a couple remaining issues with `EnumTest` related to inconsistent `pgType` coming back from `TypeInfo`.  This should be resolvable with the support of multiple `pgType` associated with a single `oid`.  The tests are altered to pass, with "TODO" markers where not yet correct.  All tests pass locally.