Stop tethering `LocalDateTime` to current zone on serialization
This removes the tethering of passed in `LocalDateTime` objects to the current time zone. Instead it tethers it to the UTC offset only for purposes of code reuse.

Tethering it to UTC at least guarantees the value won’t be changed during serialization as it sometimes is when tething to the current time zone.