feat: Add connection property to limit server error detail in exception messages
Adds a boolean connection property, LOG_SERVER_MESSAGE_DETAIL ("logServerMessage") that
controls whether the full error detail, as reported back by the server, is included in
the exception messages. This property can be used to limit potentially sensitive content
in server messages from bubbling up in exception messages and stack traces.

If set to "false" then only the top level error will be included in the exception message.
If set to "true" then all details in the server error will be included.

Default value for the new property is "true" to reflect the prior behavior of logging the
full error detail.

Closes #1577