Add 'connectSocketTimeout' to reduce socket timeout during tryConnect()
This change is connected with PR https://github.com/pgjdbc/pgjdbc/pull/1584 .
Now 'socketTimeout' property is a global query timeout, so we can't set it too low.
But on network split, it is possible that creation of a new connection
will stick on reading response from one of candidate hosts (like in PR above)
for a long time (our big socketTimeout).
The simplest solution is to decrease socket timeout during connection phase, and then restore it.