WiP: feat: allow getScale() and getPrecision() to describe unset values
A scale of null was previously reported as 0, the same as an actual 0.
This changes the value used to represent an unset scale to -127, as in the Oracle
JDBC Driver, and adds a config parameter that allows the user to
specify a different value in the connection string, to get the
old behaviour (or any other value needed).

The value returned by getPrecision() for an unset precision on a
Numeric column was also 0, which would mean a column can only store 0.
This change sets to the maximum possible value for precision
(according to the documentation and experimentation).

The reason for using a high, but possible, value for precision and
an impossible value for scale is that they have different meanings.
A high precision p means *up to* p digits, while a high scale s
means there *are* exactly s digits after the decimal point. This
difference has implications for recorded accuracy as well as
display formatting.

Signed-off-by: crwr45 <charlie.wheelerrobinson@gmail.com>