WiP: feat: allow getScale() to use a custom value for null scale
a scale of null is reported as 0. This change adds a config parameter
that allows the user to specify a new value in the connection string.

The interntion is thata value like -127 can be used, as it (probably)
is in the oracle jdbc driver, to represent a null scale on a column
when describing a ResultSet.

Signed-off-by: crwr45 <charlie.wheelerrobinson@gmail.com>