docs: update openssl key conversion example command
update openssl example command to reflect the filenames the user likely needs, and to be
consistent with the naming requirements described in the previous sentence. add -nocrypt flag to
the example command. this prevents driver from attempting to ask the user for a password for the
resulting key file (even when the input key did not have a password).