Updated openssl example command (#1763)
* docs: update openssl key conversion example command

update openssl example command to reflect the filenames the user likely needs, and to be
consistent with the naming requirements described in the previous sentence. add -nocrypt flag to
the example command. this prevents driver from attempting to ask the user for a password for the
resulting key file (even when the input key did not have a password).

* docs: add note explaining usage of sslpassword with respect to sslkey

Explain that key passwords are provided using sslpassword. Move the mention of the `-nocrypt`
flag to the note for use with password-less key files.