Round nanosecond of binary-received float8-timestamp
If a float8-timestamp is received on a binaryTransfer connection, its
nanosecond part is just truncated in toParsedTimestampBinPlain. If the
result of the subtraction just before gets slightly smaller, the
result gets smaller by 1 microsecond. Nanosecond part is rounded off
on server side so it should be handled the same way.