Allow Clobs to be based on text and Blobs on bytea
Provide a connection setting lobVarlena, which is false by default.
If true, Clobs are treated as text fields and Blobs as bytea fields,
similarly to how they are treated in MySQL and elsewhere. This makes
it easier to use the same code when dealing with multiple databases.
Comparatively few users use large objects, and this will come more
naturally to many users.