fix: send type void parameters as null value (#1922)
The previous code sent parameters of type void (used for OUT
parameters of functions) with a text value of "null".  While this
doesn't affect anything, since the value is ignored anyway, it's
somewhat confusing, especially if it shows up in debug traces and
error messages, and it's also a bit wasteful to send 4 bytes for
nothing.  So instead, send those parameters as a null value, which
seems semantically saner.