test: Refactor TestUtil.terminateBackend(...) to wait for backend process to exit
Changes terminateBackend(...) method public signature to take the connection that is being
terminated and use it's properties to create a new privileged connection to use for requesting
process termination. If terminated, wait up to the default wait time for the backend process
to actually exit before returning back to the caller. This way callers need not add their own
looping logic to ensure the backend has actually been killed as pg_terminate_backend(...) only
sends the termination signal, it does not actually wait for the process to terminate.

Also moves the CopyTest termination call outside the try-catch so errors during the cancelation
are not swallowed by the test itself.