test: Improve omni action job names and split replication to run standalone
Improves names of omni actions jobs and present them in a sorted manner so that all the
experimental jobs appear last.

Also changes the job execution so that a single job performs the replication test on their
own. Running them combined with other tests causes them to randomly fail and this is a
hopefully temporary measure to get the tests running successfully even if it is in isolation.

Gradle args as a dispatch parameter have been removed. All jobs now run test suite with code
coverage and enable Gradle build scans.

By default slow and replication tests are not run on the majority of matrix jobs. This can be
overridden with the new "Default test group" input to specify which tests they will be running:
  fast        - Only fast tests
  slow        - Fast and slow tests
  all         - All tests including replication tests