test: Add docker files and scripts to launch server built from source
Adds a docker/bin/postgres-head script that builds and launches a docker container
for a PostgreSQL server built from source. It defaults to building the latest master
branch but can be overridden by setting the PG_BRANCH_NAME environment variable.

To better take advantage of docker build step caching, the build script includes the
SHA hash of the git revision as a built argument. Repeatedly creating the container
without new commits being pushed to the upstream server source repo should not trigger
new builds as the cached image will be reused. Changes upstream will trigger a change
to to the GIT_SHA build argument and instruct docker to rebuild the image.