fix: change to updateable result set to use correctly primary or unique keys
Fix issues introduced in PR# 2199 causing problems when updating tables with PK and UK.
Originally updateable result set supports only primary keys, wich the forementioned change
and this fix unique keys are also eligible to enable updateable result set. This change will
find the better suited primary or unique key that has all the available columns in the query, but
if no match is found it will continue to fail as it used to.

Closes issue 2196