Change org.checkerframework:checker-qual to be a compileOnly dependency
Currently `checker-qual` is a Gradle implementation scope dependency and the checker-qual annotations are RUNTIME retention.

This changes to use `checker-qual-android` which has the exact same annotations as `checker-qual` but with CLASS retention. This allows us to change it to a compileOnly dependency.

Note that with checker-qual-android as a compileOnly dependency we then need to additionally explicitly add it as a testCompileOnly dependency.