fix: Provide useful error message for empty or missing passwords for SCRAM auth
Previously the driver would throw an internal error if the server requested SCRAM authentication
and no or an empty password was specified. Fixes things so that the driver provides a helpful
error message to the user that a non-empty password is required.

Also adds tests to verify the behavior and expected error messages.