Range types are unique to Postgres and is very powerful. Unfortunately driver support from pgjdbc is currently lacking and forcing users to parse and create strings.
This pull request is a first step towards full range type support in
pgjdbc. It adds support for the int4range, int8range and numrange types
by providing dedicated PGobject subclasses.
We break with the traditional design of the PGobject hierarchy and use
encapsulation through accessors instead of public mutable fields.

The types tsrange, tstzrange and daterange are currently not yet
implemented. Doing with Java 8 Date Time API types would be easy but
require Java 8. Doing them with java.sql types would be tedious and
somehow require that we get access to TimestampUtils in
PGobject.setValue which would require a redesign of the PGobject
hierarchy.

Binary mode is implemented for int4range and int8range.

Oids for int4range, int8range and numrange are registered.

Postgres treats [1,3) and (0,2] as equal for int4range and int8range
but the #equals() method in Java currently does not.

This change should not break backwards compatibility as the new range
types are subtypes of PGobject.

This a first set to have framework support for range types like
https://github.com/jOOQ/jOOQ/issues/2968