fix: Connection leak in ConnectionFactoryImpl#tryConnect #2350
When `ConnectionFactoryImpl#tryConnect` fails with an exception,
a newly created PGStream is left without closing it. The finalizer
handles the close, so it's not a problem most of the time.
But Finalizer has no guarantees on the time when it'll close
the resource. The resource may hang open for a longer period.

After the change PGStream is closed just before losing the reference.