feat: Add authenticationPluginClassName option to provide passwords at runtime
Adds authenticationPluginClassName connection property that allows end users to specify a class
that will provide the connection passwords at runtime. This allows configuring a connection with
a password that must be generated on the fly or periodically changes.

Custom implementations have access to the full connection properties and the type of authentication
that is being requested by the server, e.g. MD5 or SCRAM.