test: Disable no-arg callable statement tests in simple query mode
Disables the JDBC function call syntax callable statement tests with no arguments
when running the test suite in simple query mode. The new tests fail with simple
query mode as internal JDBC-to-native query parser generates SQL that contains
the out parameters in the SQL body which is only ignored by the server when
processed in an extended parse, bind, execute sequence. When executed as a simple
query command the server does not ignore the extra parameters and the query
fails.

This commit does not fix anything yet. It simply skips the new failing test in simple
query mode so that the rest of the quite can execute. The behavior of the driver has
not changed in the recent additions either, the new tests simply brought to light an
existing issue with simple query mode.

See #2399 for more details.