fix: use PhantomReferences instead of Obejct.finalize to track Connection leaks (#2847)
The change replaces all uses of Object.finalize with PhantomReferences.
The leaked resources (Connections) are tracked in a helper thread that is active as long as
there are connection in use. By default, the thread keeps running for 30 seconds after all
the connections are released. The timeout is set with pgjdbc.config.cleanup.thread.ttl system property.

This reverts commit 2578e8553845f8b93b9739335c3ab0d0395daf0b.

Co-authored-by: Vladimir Sitnikov <sitnikov.vladimir@gmail.com>