fix: use LocalDate.MIN/MAX, LocalDateTime.MIN/MAX, OffsetDateTime.MIN/MAX for +-infinity
The database behaviour is that it does not clamp dates exceeding low and high values
to infinity, so we should respect that in the driver as well.

In other words, if the user supplies a timestamp with year 294277 (~maximum date for timestamps), then
the driver should not convert it to infinity, and it should just pass the value to the database.

The same goes for "lowest" year which is like 4714 BC.

Previously, the driver silently converted dates before 4713 BC to -infinity,
and after the change, only LocalDate.MIN would be converted to -infinity.

Previously the dates after (LocalDateTime.MAX | OffsetDateTime.MAX)-500.nanos were clamped to infinity,
and now only MAX dates would be treated as infinity.

See https://www.postgresql.org/docs/current/datatype-datetime.html