fix: validate catalog parameter of metadata queries (pgjdbc#2947)
Report an error if the catalog parameter is not null or empty, but does
not equal the catalog the current connection applies to. This is a
breaking change.

NOTE: Even with this change, we are not 100% compatible with the JDBC
API, since it would dictate treating an empty String catalog parameter
differently from a null one. (The former should query entities that are
not associated with any catalog.) Since such entities do not exist in
PostgreSQL, returning an empty result would be ideal.
However, since the current metadata implementation treats empty
schema/table/column/etc. filters no differently from null ones,
returning empty results for queries with an empty catalog value would
make the driver's behavior inconsistent. (Adjusting how the other
filters work, on the other hand, would be a much larger breaking change.)

closes pgjdbc#2947