perf: optimize BlobOutputStream.write by buffering small consequent writes, and allocating the buffer lazily
Previously, each each blobOutput.write(byte[], ..) resulted in database call
even in the case the write size is small. It could impact cases when users
write in small chunks.
At the same time, if they already write with big chunks, we should not create
a buffer in BlobOutputStream, so now we create the buffer lazily.