feat: Add PasswordUtil and PGConnection.alterUserPassword(...) for encrypting passwords client side
Add a PasswordUtil helper that provides methods for encoding a given password client
side so that it may be used for ALTER USER and CREATE USER statement without sending
the plaintext password over the wire.

Also adds a helper to PGConnection that generates the ALTER USER command necessary to
update a user's password and executes it.

The helpers default to encoding the password using SCRAM-SHA-256 if used directly or using the
server's default password_encryption setting if used via the PGConnection helper.

Co-authored-by: Sehrope Sarkuni <sehrope@jackdb.com>
Co-authored-by: Dave Cramer <davecramer@gmail.com>