test: Force use of /dev/urandom in CI as source for java.security.egd
Force use of /dev/urandom as the source of entropy for the internal state of the secure
RNG by setting java.security.egd. This is needed to prevent a poorly configured runner
from draining its entropy by using the default of /dev/random. This is not needed on
modern kernels as they do not block after boot. However it may be required for older
kernels that will otherwise refuse to return back random bits unless they think they
have enough tracked entropy in /dev/random. Specifically, this corrects an issue with
one of the hosted CI runners that was hanging fetching bytes from SecureRandom.